requirements.txt
scikit-learn
pandas
streamlit
cohere
python-dotenv
speechrecognition
pyaudio
